GEOMESA-800 Fixed sorting null values via a FeatureStore
* Null values treated as the lowest value when using a sorting query on a FeatureStore
