+++ geomesa-core/src/test/scala/org/locationtech/geomesa/core/data/AccumuloDataStoreTest.scala
@@ -60,35 +60,7 @@ import scala.collection.JavaConversions._
 import scala.util.Random
 
 @RunWith(classOf[JUnitRunner])
-class AccumuloDataStoreTest extends Specification {
-
-  val ff = CommonFactoryFinder.getFilterFactory2
-  val geotimeAttributes = org.locationtech.geomesa.core.index.spec
-  val hints = new Hints(Hints.FEATURE_FACTORY, classOf[AvroSimpleFeatureFactory])
-  val featureFactory = CommonFactoryFinder.getFeatureFactory(hints)
-  val WGS84 = DefaultGeographicCRS.WGS84
-  val gf = JTSFactoryFinder.getGeometryFactory
-
-  val df = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss")
-  df.setTimeZone(TimeZone.getTimeZone("UTC"))
-
-  val defaultSchema = "name:String,geom:Point:srid=4326,dtg:Date"
-  val defaultGeom = WKTUtils.read("POINT(45.0 49.0)")
-  val defaultDtg = new Date(100000)
-  val defaultName = "testType"
-  val defaultFid = "fid-1"
-
-  val defaultTable = "AccumuloDataStoreTest"
-
-  val ds = DataStoreFinder.getDataStore(Map(
-    "instanceId"        -> "mycloud",
-    "zookeepers"        -> "zoo1:2181,zoo2:2181,zoo3:2181",
-    "user"              -> "myuser",
-    "password"          -> "mypassword",
-    "tableName"         -> defaultTable,
-    "useMock"           -> "true",
-    "featureEncoding"   -> "avro")).asInstanceOf[AccumuloDataStore]
-
+class AccumuloDataStoreTest extends Specification with AccumuloDataStoreDefaults {
   "AccumuloDataStore" should {
     "create a store" in {
       ds must not beNull
@@ -1239,112 +1211,6 @@ class AccumuloDataStoreTest extends Specification {
         }
         success
       }
-
     }
   }
-
-  "AccumuloFeatureStore" should {
-    "support sorting and handle time bounds" in {
-      val sftName = "sortingAndTimeBoundsTest"
-      val sft = createSchema(sftName)
-
-      val fs = ds.getFeatureSource(sftName).asInstanceOf[SimpleFeatureStore]
-      fs.getQueryCapabilities.supportsSorting(Array(SortBy.NATURAL_ORDER)) must beTrue
-      fs.getQueryCapabilities.supportsSorting(Array(ff.sort("dtg", SortOrder.ASCENDING))) must beTrue
-
-      val defaultInterval = ds.getTimeBounds(sft.getTypeName)
-      defaultInterval.getStartMillis must be equalTo 0
-
-      val sfBuilder = new SimpleFeatureBuilder(sft)
-      val date1 = new DateTime("2014-01-02").toDate
-      sfBuilder.addAll(List("johndoe", gf.createPoint(new Coordinate(0, 0)), date1))
-      val f1 = sfBuilder.buildFeature("f1")
-
-      fs.addFeatures(DataUtilities.collection(List(f1)))
-
-      val secondInterval = ds.getTimeBounds(sft.getTypeName)
-      secondInterval.getStartMillis must be equalTo date1.getTime
-      secondInterval.getEndMillis must be equalTo date1.getTime
-
-      val date2 = new DateTime("2014-01-03").toDate
-      sfBuilder.addAll(List("johndoe", gf.createPoint(new Coordinate(0, 0)), date2))
-      val f2 = sfBuilder.buildFeature("f2")
-
-      val date3 = new DateTime("2014-01-01").toDate
-      sfBuilder.addAll(List("johndoe", gf.createPoint(new Coordinate(0, 0)), date3))
-      val f3 = sfBuilder.buildFeature("f3")
-
-      fs.addFeatures(DataUtilities.collection(List(f2, f3)))
-
-      val thirdInterval = ds.getTimeBounds(sft.getTypeName)
-      thirdInterval.getStartMillis must be equalTo date3.getTime
-      thirdInterval.getEndMillis must be equalTo date2.getTime
-
-      "ascending on date" >> {
-        val dtgAscendingQ = new Query("test", Filter.INCLUDE)
-        dtgAscendingQ.setSortBy(Array(ff.sort("dtg", SortOrder.ASCENDING)))
-
-        val features = fs.getFeatures(dtgAscendingQ).features().toIterator
-        val res: Seq[Long] = features.map(_.getAttribute("dtg").asInstanceOf[Date].getTime).toSeq
-        res must beSorted
-        res.head must beLessThan(res(1))
-      }
-
-      "descending on date" >> {
-        val dtgDescending = new Query("test", Filter.INCLUDE)
-        dtgDescending.setSortBy(Array(ff.sort("dtg", SortOrder.DESCENDING)))
-
-        val features = fs.getFeatures(dtgDescending).features().toIterator
-        val res: Seq[Long] = features.map(_.getAttribute("dtg").asInstanceOf[Date].getTime).toSeq
-        res.reverse must beSorted
-        res.head must beGreaterThan(res(1))
-      }
-    }
-  }
-
-  /**
-   * Create a schema. Schema name should be unique, otherwise tests will interfere with each other.
-   *
-   * @param sftName
-   * @param spec
-   * @param dateField
-   */
-  def createSchema(sftName: String,
-                   spec: String = defaultSchema,
-                   dateField: Option[String] = Some("dtg"),
-                   dataStore: DataStore = ds) = {
-    val sft = SimpleFeatureTypes.createType(sftName, spec)
-    dateField.foreach(dt => sft.getUserData.put(SF_PROPERTY_START_TIME, dt))
-    dataStore.createSchema(sft)
-    sft
-  }
-
-  def addDefaultPoint(sft: SimpleFeatureType,
-                      attributes: List[AnyRef] = List(defaultName, defaultGeom, defaultDtg),
-                      fid: String = defaultFid,
-                      dataStore: DataStore = ds) = {
-    val fs = dataStore.getFeatureSource(sft.getTypeName).asInstanceOf[AccumuloFeatureStore]
-
-    // create a feature
-    val builder = new SimpleFeatureBuilder(sft, featureFactory)
-    builder.addAll(attributes)
-    val liveFeature = builder.buildFeature(fid)
-
-    // make sure we ask the system to re-use the provided feature-ID
-    liveFeature.getUserData.put(Hints.USE_PROVIDED_FID, java.lang.Boolean.TRUE)
-    val featureCollection = new DefaultFeatureCollection(sft.getTypeName, sft)
-    featureCollection.add(liveFeature)
-    fs.addFeatures(featureCollection)
-  }
-
-  def createTestFeatures(sft: SimpleFeatureType) = (0 until 6).map { i =>
-    val builder = new SimpleFeatureBuilder(sft, featureFactory)
-    builder.set("geom", WKTUtils.read("POINT(45.0 45.0)"))
-    builder.set("dtg", "2012-01-02T05:06:07.000Z")
-    builder.set("name",i.toString)
-    val sf = builder.buildFeature(i.toString)
-    sf.getUserData()(Hints.USE_PROVIDED_FID) = java.lang.Boolean.TRUE
-    sf
-  }
-
 }
